adding WidgetParams to its own module
